Revert "added jQuery.noConflict support."
This reverts commit b648918d3cdfca38af5333be0c5d27b7ae45a687.

Michael's code is not "wrong" It just pushes the chore of maintenance of globals into individual source files.
Found wrapping minified build into jQuery.NoConflict-compatible wrapper a simpler solution from the stand-point of maintenance.

In other words, while the actual patch is reverted, the idea is incorporated and moved up the chain to build-time.